moved layouts to corresponding app folders and moved infobox into guidelines folder
