--- src/components/InfoBox/index.js
@@ -1,3 +1,0 @@
-import InfoBox from "./InfoBox";
-
-export default InfoBox;
